refactor(desktop): extract workspace init to dedicated module with cleaner patterns
- Extract initializeWorkspaceWorktree to utils/init-worktree.ts (~290 lines moved)
- Create context object that centralizes cancellation checks and cleanup
- Remove duplicated cleanup logic (was repeated 3 times in cancellation checks)
- Simplify main flow with early returns and helper functions
- Split into focused step functions for better readability

Benefits:
- workspaces.ts reduced from ~1900 to ~1600 lines
- Cancellation + cleanup logic is now in one place
- Each init step is a focused, testable function
- Easier to add new steps or modify existing ones
